ENH Add optional usage of exp/run params for submit_slurm. BUG cache file undefined.

+++ launch_scripts/launch_airflow.py
@@ -138,8 +138,8 @@ if __name__ == "__main__":
     use_kerberos: bool = (
         True  # Always copy kerberos ticket so non-active experiments can work.
     )
+    cache_file: Optional[str] = os.getenv("KRB5CCNAME")
     if os.getenv("Authorization") is None:
-        cache_file: Optional[str] = os.getenv("KRB5CCNAME")
         if cache_file is None:
             logger.info("No Kerberos cache. Try running `kinit` and resubmitting.")
             sys.exit(-1)
@@ -378,8 +378,9 @@ if __name__ == "__main__":
         logger.debug("Removing duplicate Kerberos credentials.")
         # This should be defined if we get here
         # Format is FILE:/.../...
-        os.remove(cache_file[5:])
-        os.rmdir(f"{os.path.expanduser('~')}/.tmp_cache")
+        if cache_file is not None:
+            os.remove(cache_file[5:])
+            os.rmdir(f"{os.path.expanduser('~')}/.tmp_cache")
 
     if dag_state == "failed":
         sys.exit(1)
